Moving code into function_wrapper.rs.
No intentional functional changes.
