+++ engine/src/lib.rs
@@ -15,6 +15,7 @@
 mod additional_cpp_generator;
 mod bridge_converter;
 mod byvalue_checker;
+mod function_wrapper;
 mod known_types;
 mod namespace_organizer;
 mod parse;
